--- example/.flutter-plugins-dependencies
@@ -1,1 +1,0 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"path_provider","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/path_provider-2.0.1/","dependencies":[]},{"name":"url_launcher","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/url_launcher-5.3.0/","dependencies":[]}],"android":[{"name":"path_provider","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/path_provider-2.0.1/","dependencies":[]},{"name":"url_launcher","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/url_launcher-5.3.0/","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/path_provider_macos-2.0.0/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/path_provider_linux-2.0.0/","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/path_provider_windows-2.0.0/","dependencies":[]}],"web":[{"name":"url_launcher_web","path":"/Users/zmt/Documents/flutter/stable/.pub-cache/hosted/pub.flutter-io.cn/url_launcher_web-0.1.5+3/","dependencies":[]}]},"dependencyGraph":[{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux","path_provider_windows"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]},{"name":"url_launcher","dependencies":["url_launcher_web"]},{"name":"url_launcher_web","dependencies":[]}],"date_created":"2021-03-08 20:01:31.691145","version":"2.0.0"}
