--- example/ios/Flutter/flutter_export_environment.sh
@@ -1,14 +1,0 @@
-#!/bin/sh
-# This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/zmt/Documents/flutter/stable"
-export "FLUTTER_APPLICATION_PATH=/Users/zmt/Documents/github/extended_nested_scroll_view/example"
-export "FLUTTER_TARGET=/Users/zmt/Documents/github/extended_nested_scroll_view/example/lib/main.dart"
-export "FLUTTER_BUILD_DIR=build"
-export "SYMROOT=${SOURCE_ROOT}/../build/ios"
-export "FLUTTER_BUILD_NAME=1.0.0"
-export "FLUTTER_BUILD_NUMBER=1"
-export "DART_DEFINES=flutter.inspector.structuredErrors%3Dtrue"
-export "DART_OBFUSCATION=false"
-export "TRACK_WIDGET_CREATION=true"
-export "TREE_SHAKE_ICONS=false"
-export "PACKAGE_CONFIG=/Users/zmt/Documents/github/extended_nested_scroll_view/example/.dart_tool/package_config.json"
