+++ example/lib/pages/complex/scroll_to_top.dart
@@ -1,6 +1,5 @@
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
-import 'package:flutter/material.dart'
-    hide NestedScrollView, NestedScrollViewState;
+import 'package:flutter/material.dart';
 import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
 import 'package:loading_more_list/loading_more_list.dart';
 
@@ -22,8 +21,8 @@ class ScrollToTopDemo extends StatefulWidget {
 class _ScrollToTopDemoState extends State<ScrollToTopDemo>
     with TickerProviderStateMixin {
   TabController primaryTC;
-  final GlobalKey<NestedScrollViewState> _key =
-      GlobalKey<NestedScrollViewState>();
+  final GlobalKey<ExtendedNestedScrollViewState> _key =
+      GlobalKey<ExtendedNestedScrollViewState>();
   @override
   void initState() {
     primaryTC = TabController(length: 2, vsync: this);
@@ -44,7 +43,7 @@ class _ScrollToTopDemoState extends State<ScrollToTopDemo>
         child: const Icon(Icons.file_upload),
         onPressed: () {
           ///scroll current tab list
-          _key.currentState.currentInnerPosition.animateTo(0.0,
+          _key.currentState.outerController.animateTo(0.0,
               duration: const Duration(seconds: 1), curve: Curves.easeIn);
 
           ///scroll all tab list
@@ -64,7 +63,7 @@ class _ScrollToTopDemoState extends State<ScrollToTopDemo>
         statusBarHeight +
             //pinned SliverAppBar height in header
             kToolbarHeight;
-    return NestedScrollView(
+    return ExtendedNestedScrollView(
       key: _key,
       headerSliverBuilder: (BuildContext c, bool f) {
         return <Widget>[
@@ -86,13 +85,7 @@ class _ScrollToTopDemoState extends State<ScrollToTopDemo>
         return pinnedHeaderHeight;
       },
       //2.[inner scrollables in tabview sync issue](https://github.com/flutter/flutter/issues/21868)
-      innerScrollPositionKeyBuilder: () {
-        String index = 'Tab';
-
-        index += primaryTC.index.toString();
-
-        return Key(index);
-      },
+      onlyOneScrollInBody: true,
       body: Column(
         children: <Widget>[
           TabBar(
@@ -112,46 +105,40 @@ class _ScrollToTopDemoState extends State<ScrollToTopDemo>
             child: TabBarView(
               controller: primaryTC,
               children: <Widget>[
-                NestedScrollViewInnerScrollPositionKeyWidget(
-                  const Key('Tab0'),
-                  GlowNotificationWidget(
-                    ListView.builder(
-                      //store Page state
-                      key: const PageStorageKey<String>('Tab0'),
-                      physics: const ClampingScrollPhysics(),
-                      itemBuilder: (BuildContext c, int i) {
-                        return Container(
-                          alignment: Alignment.center,
-                          height: 60.0,
-                          child: Text(
-                              const Key('Tab0').toString() + ': ListView$i'),
-                        );
-                      },
-                      itemCount: 50,
-                    ),
-                    showGlowLeading: false,
+                GlowNotificationWidget(
+                  ListView.builder(
+                    //store Page state
+                    key: const PageStorageKey<String>('Tab0'),
+                    physics: const ClampingScrollPhysics(),
+                    itemBuilder: (BuildContext c, int i) {
+                      return Container(
+                        alignment: Alignment.center,
+                        height: 60.0,
+                        child:
+                            Text(const Key('Tab0').toString() + ': ListView$i'),
+                      );
+                    },
+                    itemCount: 50,
                   ),
+                  showGlowLeading: false,
                 ),
-                NestedScrollViewInnerScrollPositionKeyWidget(
-                  const Key('Tab1'),
-                  GlowNotificationWidget(
-                    ListView.builder(
-                      //store Page state
-                      key: const PageStorageKey<String>('Tab1'),
-                      physics: const ClampingScrollPhysics(),
-                      itemBuilder: (BuildContext c, int i) {
-                        return Container(
-                          alignment: Alignment.center,
-                          height: 60.0,
-                          child: Text(
-                              const Key('Tab1').toString() + ': ListView$i'),
-                        );
-                      },
-                      itemCount: 50,
-                    ),
-                    showGlowLeading: false,
+                GlowNotificationWidget(
+                  ListView.builder(
+                    //store Page state
+                    key: const PageStorageKey<String>('Tab1'),
+                    physics: const ClampingScrollPhysics(),
+                    itemBuilder: (BuildContext c, int i) {
+                      return Container(
+                        alignment: Alignment.center,
+                        height: 60.0,
+                        child:
+                            Text(const Key('Tab1').toString() + ': ListView$i'),
+                      );
+                    },
+                    itemCount: 50,
                   ),
-                )
+                  showGlowLeading: false,
+                ),
               ],
             ),
           )
