+++ example/lib/pages/simple/dynamic_pinned_header_height.dart
@@ -1,5 +1,5 @@
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
-import 'package:flutter/material.dart' hide NestedScrollView;
+import 'package:flutter/material.dart';
 import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
 import 'package:loading_more_list/loading_more_list.dart';
 
@@ -60,7 +60,7 @@ class _DynamicPinnedHeaderHeightDemoState
         statusBarHeight +
             //pinned SliverAppBar height in header
             kToolbarHeight;
-    return NestedScrollView(
+    return ExtendedNestedScrollView(
       controller: sc,
       headerSliverBuilder: (BuildContext c, bool f) {
         return <Widget>[
@@ -84,13 +84,7 @@ class _DynamicPinnedHeaderHeightDemoState
         return pinnedHeaderHeight;
       },
       //2.[inner scrollables in tabview sync issue](https://github.com/flutter/flutter/issues/21868)
-      innerScrollPositionKeyBuilder: () {
-        String index = 'Tab';
-
-        index += primaryTC.index.toString();
-
-        return Key(index);
-      },
+      onlyOneScrollInBody: true,
       body: Column(
         children: <Widget>[
           TabBar(
@@ -110,46 +104,40 @@ class _DynamicPinnedHeaderHeightDemoState
             child: TabBarView(
               controller: primaryTC,
               children: <Widget>[
-                NestedScrollViewInnerScrollPositionKeyWidget(
-                  const Key('Tab0'),
-                  GlowNotificationWidget(
-                    ListView.builder(
-                      //store Page state
-                      key: const PageStorageKey<String>('Tab0'),
-                      physics: const ClampingScrollPhysics(),
-                      itemBuilder: (BuildContext c, int i) {
-                        return Container(
-                          alignment: Alignment.center,
-                          height: 60.0,
-                          child: Text(
-                              const Key('Tab0').toString() + ': ListView$i'),
-                        );
-                      },
-                      itemCount: 50,
-                    ),
-                    showGlowLeading: false,
+                GlowNotificationWidget(
+                  ListView.builder(
+                    //store Page state
+                    key: const PageStorageKey<String>('Tab0'),
+                    physics: const ClampingScrollPhysics(),
+                    itemBuilder: (BuildContext c, int i) {
+                      return Container(
+                        alignment: Alignment.center,
+                        height: 60.0,
+                        child:
+                            Text(const Key('Tab0').toString() + ': ListView$i'),
+                      );
+                    },
+                    itemCount: 50,
                   ),
+                  showGlowLeading: false,
                 ),
-                NestedScrollViewInnerScrollPositionKeyWidget(
-                  const Key('Tab1'),
-                  GlowNotificationWidget(
-                    ListView.builder(
-                      //store Page state
-                      key: const PageStorageKey<String>('Tab1'),
-                      physics: const ClampingScrollPhysics(),
-                      itemBuilder: (BuildContext c, int i) {
-                        return Container(
-                          alignment: Alignment.center,
-                          height: 60.0,
-                          child: Text(
-                              const Key('Tab1').toString() + ': ListView$i'),
-                        );
-                      },
-                      itemCount: 50,
-                    ),
-                    showGlowLeading: false,
+                GlowNotificationWidget(
+                  ListView.builder(
+                    //store Page state
+                    key: const PageStorageKey<String>('Tab1'),
+                    physics: const ClampingScrollPhysics(),
+                    itemBuilder: (BuildContext c, int i) {
+                      return Container(
+                        alignment: Alignment.center,
+                        height: 60.0,
+                        child:
+                            Text(const Key('Tab1').toString() + ': ListView$i'),
+                      );
+                    },
+                    itemCount: 50,
                   ),
-                )
+                  showGlowLeading: false,
+                ),
               ],
             ),
           )
