+++ example/lib/pages/simple/extened_nested_scroll_view_demo.dart
@@ -1,7 +1,8 @@
 import 'package:example/common/common.dart';
 import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
-import 'package:flutter/material.dart' hide NestedScrollView;
+import 'package:flutter/material.dart';
 import 'package:extended_nested_scroll_view/extended_nested_scroll_view.dart';
+import 'package:flutter_swiper/flutter_swiper.dart';
 import 'package:loading_more_list/loading_more_list.dart';
 
 @FFRoute(
@@ -13,42 +14,29 @@ import 'package:loading_more_list/loading_more_list.dart';
     'order': 0,
   },
 )
-class OldExtendedNestedScrollViewDemo extends StatefulWidget {
+class ExtendedNestedScrollViewDemo extends StatefulWidget {
   @override
-  _OldExtendedNestedScrollViewDemoState createState() =>
-      _OldExtendedNestedScrollViewDemoState();
+  _ExtendedNestedScrollViewDemoState createState() =>
+      _ExtendedNestedScrollViewDemoState();
 }
 
-class _OldExtendedNestedScrollViewDemoState
-    extends State<OldExtendedNestedScrollViewDemo>
-    with TickerProviderStateMixin {
+class _ExtendedNestedScrollViewDemoState
+    extends State<ExtendedNestedScrollViewDemo> with TickerProviderStateMixin {
   TabController primaryTC;
   TabController secondaryTC;
 
   @override
   void initState() {
     primaryTC = TabController(length: 2, vsync: this);
-    primaryTC.addListener(tabControlerListener);
-    secondaryTC = TabController(length: 4, vsync: this);
+    secondaryTC = TabController(length: 5, vsync: this);
     super.initState();
   }
 
   @override
   void dispose() {
-    primaryTC.removeListener(tabControlerListener);
     primaryTC.dispose();
     secondaryTC.dispose();
     super.dispose();
-  }
-
-  //when primary tabcontroller tab,rebuild headerSliverBuilder
-  //click fire twice (due to animation),gesture fire onetime
-  int index;
-  void tabControlerListener() {
-    if (index != primaryTC.index) {
-      //your code
-      index = primaryTC.index;
-    }
   }
 
   @override
@@ -65,7 +53,7 @@ class _OldExtendedNestedScrollViewDemoState
         statusBarHeight +
             //pinned SliverAppBar height in header
             kToolbarHeight;
-    return NestedScrollView(
+    return ExtendedNestedScrollView(
       headerSliverBuilder: (BuildContext c, bool f) {
         return buildSliverHeader();
       },
@@ -74,17 +62,23 @@ class _OldExtendedNestedScrollViewDemoState
         return pinnedHeaderHeight;
       },
       //2.[inner scrollables in tabview sync issue](https://github.com/flutter/flutter/issues/21868)
-      innerScrollPositionKeyBuilder: () {
-        String index = 'Tab';
-        if (primaryTC.index == 0) {
-          index += primaryTC.index.toString() + secondaryTC.index.toString();
-        } else {
-          index += primaryTC.index.toString();
-        }
-        return Key(index);
-      },
+      onlyOneScrollInBody: true,
       body: Column(
         children: <Widget>[
+          SizedBox(
+            height: 150,
+            child: Swiper(
+              itemBuilder: (BuildContext context, int index) {
+                return Image.network(
+                  'http://via.placeholder.com/350x150',
+                  fit: BoxFit.fill,
+                );
+              },
+              itemCount: 3,
+              pagination: const SwiperPagination(),
+              autoplay: true,
+            ),
+          ),
           TabBar(
             controller: primaryTC,
             labelColor: Colors.blue,
@@ -102,27 +96,24 @@ class _OldExtendedNestedScrollViewDemoState
             child: TabBarView(
               controller: primaryTC,
               children: <Widget>[
-                SecondaryTabView('Tab0', secondaryTC, true),
-                NestedScrollViewInnerScrollPositionKeyWidget(
-                  const Key('Tab1'),
-                  GlowNotificationWidget(
-                    ListView.builder(
-                      //store Page state
-                      key: const PageStorageKey<String>('Tab1'),
-                      physics: const ClampingScrollPhysics(),
-                      itemBuilder: (BuildContext c, int i) {
-                        return Container(
-                          alignment: Alignment.center,
-                          height: 60.0,
-                          child: Text(
-                              const Key('Tab1').toString() + ': ListView$i'),
-                        );
-                      },
-                      itemCount: 50,
-                    ),
-                    showGlowLeading: false,
+                SecondaryTabView('Tab0', secondaryTC),
+                GlowNotificationWidget(
+                  ListView.builder(
+                    //store Page state
+                    key: const PageStorageKey<String>('Tab1'),
+                    physics: const ClampingScrollPhysics(),
+                    itemBuilder: (BuildContext c, int i) {
+                      return Container(
+                        alignment: Alignment.center,
+                        height: 60.0,
+                        child:
+                            Text(const Key('Tab1').toString() + ': ListView$i'),
+                      );
+                    },
+                    itemCount: 50,
                   ),
-                )
+                  showGlowLeading: false,
+                ),
               ],
             ),
           )
