+++ example/pubspec.yaml
@@ -24,10 +24,12 @@ dependencies:
   ff_annotation_route_library: ^2.0.0
   flutter:
     sdk: flutter  
+  flutter_swiper: any    
   intl: ^0.15.7  
   loading_more_list: any
   pull_to_refresh_notification: any
   url_launcher: 5.3.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
