+++ lib/extended_nested_scroll_view.dart
@@ -1,6 +1,3 @@
 library extended_nested_scroll_view;
 
-//export 'src/extended_nested_scroll_view.dart';
-export 'src/nested_scroll_view_inner_scroll_position_key_widget.dart';
-export 'src/old_extended_nested_scroll_view.dart';
-export 'src/util.dart';
+export 'src/extended_nested_scroll_view.dart';
