--- lib/src/util.dart
@@ -1,2 +1,0 @@
-///get type from T
-Type typeOf<T>() => T;
