+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: extended_nested_scroll_view
 description: extended nested scroll view to fix pinned header and inner scrollables sync issues.
-version: 3.0.1
+version: 4.0.0
 homepage: https://github.com/fluttercandies/extended_nested_scroll_view
 
 environment:
